// packai/frontend/types/packing.ts
export interface ContainerInput {
  name?: string;
  length: number;
  width: number;
  height: number;
}

export interface PackageInput {
  id?: string;
  name: string;
  length: number;
  width: number;
  height: number;
  quantity: number;
  color?: string;
}

export interface Placement {
  packageTypeId: string;
  packageName: string;
  x: number;
  y: number;
  z: number;
  length: number;
  width: number;
  height: number;
  rotation: [number, number, number];
  color?: string;
}

export interface PackedContainer {
  id: number;
  name: string;
  length: number;
  width: number;
  height: number;
  placements: Placement[];
  usedVolume: number;
  emptyVolume: number;
  utilization: number;
}

export interface OptimizationSummary {
  containersRequired: number;
  averageUtilization: number;
  emptySpacePercentage: number;
  totalPackages: number;
  packedPackages: number;
}

export interface OptimizationResult {
  summary: OptimizationSummary;
  containers: PackedContainer[];
  aiInsight: string;
}

export interface HistoricalRun {
  id: string;
  containerCount: number;
  averageUtilization: number;
  emptySpacePercentage: number;
  inputData: {
    container: ContainerInput;
    packages: PackageInput[];
  };
  resultData: OptimizationResult;
  createdAt: string;
}